Move absolute hydration code to tmd.fe.absolute.hydration
* Provides a namespace for absolute related code
